add State to somethingGoodForm
